fix: fix type paths
