+++ packages/vchart/src/chart/common/interface.ts
@@ -24,7 +24,8 @@ import type { ISunburstSeriesSpec } from '../../series/sunburst/interface';
 import type { ITreemapSeriesSpec } from '../../series/treemap/interface';
 import type { IWaterfallSeriesSpec } from '../../series/waterfall/interface';
 import type { ICorrelationSeriesSpec } from '../../series/correlation/interface';
-import type { ICartesianAxisSpec, IPolarAxisSpec } from '../../component';
+import type { ICartesianAxisSpec } from '../../component/axis/cartesian/interface';
+import type { IPolarAxisSpec } from '../../component/axis/polar/interface';
 import type { ICartesianCrosshairSpec, IPolarCrosshairSpec } from '../../component/crosshair/interface';
 import type { IMarkLineSpec } from '../../component/marker/mark-line/interface';
 import type { IMarkAreaSpec } from '../../component/marker/mark-area/interface';
