+++ packages/vchart/src/compile/mark/interface.ts
@@ -1,9 +1,6 @@
 import type { IMarkStateStyle, MarkType } from '../../mark/interface';
 import type { IModel } from '../../model/interface';
-import type { GrammarItemCompileOption, GrammarItemInitOption } from '../interface';
-// eslint-disable-next-line no-duplicate-imports
-import type { IGrammarItem } from '../interface';
-import type { MarkStateManager } from './mark-state-manager';
+import type { GrammarItemCompileOption, GrammarItemInitOption, IGrammarItem } from '../interface';
 import type { DataView } from '@visactor/vdataset';
 import type {
   IAnimate,
@@ -19,6 +16,21 @@ import type {
 import type { Maybe, Datum, StringOrNumber } from '../../typings';
 import type { IRegion } from '../../region/interface';
 import type { ICompilableData } from '../data/interface';
+
+export interface IMarkStateManager {
+  getStateInfoList: () => IStateInfo[];
+  getStateInfo: (stateValue: StateValue) => IStateInfo;
+  addStateInfo: (stateInfo: IStateInfo) => void;
+  changeStateInfo: (stateInfo: Partial<IStateInfo>) => void;
+  clearStateInfo: (stateValues: StateValue[]) => void;
+  checkOneState: (
+    renderNode: IElement,
+    datum: Datum | Datum[],
+    state: IStateInfo,
+    isMultiMark?: boolean
+  ) => 'in' | 'out' | 'skip';
+  checkState: (renderNode: IElement, datum: Datum | Datum[]) => StateValue[];
+}
 
 export interface IMarkData extends ICompilableData {
   setCompiledProductId: (name: string) => any;
@@ -57,7 +69,7 @@ export interface ICompilableMark extends IGrammarItem {
   setDataView: (d?: DataView, productId?: string) => void;
 
   // 状态
-  state: MarkStateManager;
+  state: IMarkStateManager;
   readonly stateStyle: IMarkStateStyle<any>;
   hasState: (state: string) => boolean;
   getState: (state: string) => any;
