+++ packages/vchart/src/component/data-zoom/scroll-bar/interface.ts
@@ -2,8 +2,7 @@ import type { ILayoutItemSpec } from '../../../layout/interface';
 import type { IPadding } from '../../../typings';
 import type { IRectMarkSpec } from '../../../typings/visual';
 import type { ComponentThemeWithDirection } from '../../interface';
-import type { IFilterMode } from '../interface';
-import type { IDataFilterComponentSpec } from '../interface';
+import type { IFilterMode, IDataFilterComponentSpec } from '../interface';
 
 export interface IScrollBarStyle {
   /**
