+++ packages/vchart/src/component/interface/theme.ts
@@ -4,7 +4,8 @@ import type { ICrosshairTheme } from '../crosshair/interface';
 import type { IDataZoomTheme } from '../data-zoom/data-zoom/interface';
 import type { IScrollBarTheme } from '../data-zoom/scroll-bar/interface';
 import type { IIndicatorTheme } from '../indicator/interface';
-import type { IDiscreteLegendTheme, IColorLegendTheme, ISizeLegendTheme } from '../legend';
+import type { IDiscreteLegendTheme } from '../legend/discrete/interface';
+import type { IColorLegendTheme, ISizeLegendTheme } from '../legend/continuous/interface';
 import type { IMapLabelTheme } from '../map-label/interface';
 import type { IMarkAreaTheme } from '../marker/mark-area/interface';
 import type { IMarkLineTheme } from '../marker/mark-line/interface';
