+++ packages/vchart/src/data/transforms/aggregation.ts
@@ -1,60 +1,17 @@
 import type { DataView } from '@visactor/vdataset';
-import type {
-  IAggrType,
-  ICoordinateOption,
-  IDataPointSpec,
-  IDataPos,
-  IMarkerSupportSeries
-} from '../../component/marker/interface';
-import type { CoordinateType, Datum, StringOrNumber } from '../../typings';
-
+import type { IAggrType, IDataPointSpec, IDataPos, IMarkerSupportSeries } from '../../component/marker/interface';
+import type { Datum, StringOrNumber } from '../../typings';
 import { array, isArray, isFunction, isPlainObject, isString, isValid } from '@visactor/vutils';
 import { variance, average, min, max, sum, standardDeviation, median } from '../../util/math';
 import { isAggrSpec } from '../../component/marker/utils';
-
-export type IOption = {
-  field: string;
-};
-
-export type IOptionAggrField = {
-  field: string;
-  aggrType: IAggrType;
-};
-
-export type IOptionPos = IOptionAggrField | string | number | StringOrNumber[];
-
-export type IOptionSeries = {
-  getRelativeSeries: () => IMarkerSupportSeries;
-  getStartRelativeSeries: () => IMarkerSupportSeries;
-  getEndRelativeSeries: () => IMarkerSupportSeries;
-};
-
-export type IOptionCallback = (
-  relativeSeriesData: any,
-  startRelativeSeriesData: any,
-  endRelativeSeriesData: any,
-  relativeSeries: IMarkerSupportSeries,
-  startRelative: IMarkerSupportSeries,
-  endRelative: IMarkerSupportSeries
-) => IOptionPos;
-
-export type IOptionAggr = {
-  x?: IOptionPos | IOptionCallback;
-  y?: IOptionPos | IOptionCallback;
-  angle?: IOptionPos | IOptionCallback;
-  radius?: IOptionPos | IOptionCallback;
-  areaName?: string | IOptionCallback;
-  getRefRelativeSeries?: () => IMarkerSupportSeries;
-} & IOptionSeries;
-
-export type IOptionWithCoordinates = {
-  coordinates:
-    | ((seriesData: Datum[], relativeSeries: IMarkerSupportSeries) => IDataPointSpec[] | IDataPointSpec)
-    | IDataPointSpec
-    | IDataPointSpec[];
-  getSeriesByIdOrIndex: (seriesUserId: StringOrNumber, seriesIndex: number) => IMarkerSupportSeries;
-  coordinateType: CoordinateType;
-} & IOptionSeries;
+import type {
+  IOption,
+  IOptionAggr,
+  IOptionAggrField,
+  IOptionCallback,
+  IOptionPos,
+  IOptionWithCoordinates
+} from './interface';
 
 export const markerMin = (_data: Array<DataView>, opt: IOption) => {
   const data = _data[0].latestData as Datum[];
