+++ packages/vchart/src/series/correlation/animation.ts
@@ -1,14 +1,7 @@
 import type { IAnimationTypeConfig } from '@visactor/vgrammar-core';
 import { Factory } from '../../core/factory';
 import { ScaleInOutAnimation } from '../../animation/config';
-
-export type CorrelationMarks = 'point' | 'label';
-
-export type CorrelationAppearPreset = 'scaleIn' | 'fadeIn';
-
-export interface ICorrelationAnimationParams {
-  [key: string]: object;
-}
+import type { CorrelationAppearPreset, ICorrelationAnimationParams } from './interface';
 
 export const correlationPresetAnimation = (
   _params: ICorrelationAnimationParams, // 此处仅为保持结构统一
