+++ packages/vchart/src/series/line/line.ts
@@ -7,9 +7,8 @@ import { mixin } from '@visactor/vutils';
 import type { Datum } from '../../typings';
 import { animationConfig, userAnimationConfig } from '../../animation/utils';
 import { registerLineAnimation, registerScaleInOutAnimation } from '../../animation/config';
-import type { ILineSeriesSpec } from './interface';
+import type { ILineSeriesSpec, LineAppearPreset } from './interface';
 import type { IStateAnimateSpec } from '../../animation/spec';
-import type { LineAppearPreset } from './animation';
 import { lineSeriesMark } from './constant';
 import { registerLineMark } from '../../mark/line';
 import { registerSymbolMark } from '../../mark/symbol';
