+++ packages/vchart/src/series/liquid/liquid.ts
@@ -4,9 +4,8 @@ import { SeriesMarkNameEnum, SeriesTypeEnum } from '../interface/type';
 import { isValid, max } from '@visactor/vutils';
 import type { Datum, ILiquidMarkSpec, IPoint } from '../../typings';
 import { animationConfig, userAnimationConfig } from '../../animation/utils';
-import type { ILiquidPadding, ILiquidSeriesSpec, LiquidShapeType } from './interface';
+import type { ILiquidPadding, ILiquidSeriesSpec, LiquidAppearPreset, LiquidShapeType } from './interface';
 import type { IStateAnimateSpec } from '../../animation/spec';
-import type { LiquidAppearPreset } from './animation';
 // eslint-disable-next-line no-duplicate-imports
 import { registerLiquidAnimation } from './animation';
 import { LiquidSeriesMark } from './constant';
