+++ packages/vchart/src/series/rose/animation.ts
@@ -1,13 +1,7 @@
 import type { IAnimationTypeConfig } from '@visactor/vgrammar-core';
 import type { Datum } from '../../typings';
 import { Factory } from '../../core/factory';
-
-export interface IRoseAnimationParams {
-  innerRadius: () => number;
-  growField?: 'angle' | 'radius';
-}
-
-export type RoseAppearPreset = 'growAngle' | 'growRadius' | 'fadeIn';
+import type { IRoseAnimationParams, RoseAppearPreset } from './interface';
 
 export const Appear_Grow = (params: IRoseAnimationParams): IAnimationTypeConfig => {
   const from = params.growField === 'angle' ? 0 : params.innerRadius;
